Improve pps_document_fonts_scan API
The "return" value means only whether fonts were found. This is a bug
from previously in Evince and Poppler[1]. So don't return anything,
and instead always scan the whole document, as that's how it makes
most sense to use it. This allows to keep all the logic more contained
in pdf_document_fonts_scan, and to require less boilerplate around.

[1] https://gitlab.freedesktop.org/poppler/poppler/-/merge_requests/1487